fix(Color): cast to int to call different method and prevent StackOverflow (#42)
* fix(Color): cast to int to call different method and prevent StackOverflow

* chore: update copyright

+++ nui/src/main/java/org/terasology/nui/Color.java
@@ -1,18 +1,5 @@
-/*
- * Copyright 2014 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *  http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+// Copyright 2021 The Terasology Foundation
+// SPDX-License-Identifier: Apache-2.0
 package org.terasology.nui;
 
 import com.google.common.base.Preconditions;
@@ -263,7 +250,7 @@ public class Color implements Colorc {
      * @return this
      */
     public Color setRed(float value) {
-        return setRed(value * MAX);
+        return setRed((int) value * MAX);
     }
 
     /**
@@ -284,7 +271,7 @@ public class Color implements Colorc {
      * @return this
      */
     public Color setGreen(float value) {
-        return setGreen(value * MAX);
+        return setGreen((int) value * MAX);
     }
 
 
@@ -305,7 +292,7 @@ public class Color implements Colorc {
      * @return this
      */
     public Color setBlue(float value) {
-        return setBlue(value * MAX);
+        return setBlue((int) value * MAX);
     }
 
     /**
@@ -325,7 +312,7 @@ public class Color implements Colorc {
      * @return this
      */
     public Color setAlpha(float value) {
-        return setAlpha(value * MAX);
+        return setAlpha((int) value * MAX);
     }
